refactor(runner): the file is named for what it builds
`json-utils.ts` held one JSON thing, and it left in the preceding change.
What remains builds an encodable form: the alias-binding walk, the two
serializers, and the internal graph serializer. The name says so now, as
does the test file beside it and its top-level `describe`.

Renamed with `git mv`, so the history follows. Ten importers, four live
documents and one pattern README follow the path; `docs/history/` keeps
the old name, since a historical record describes the tree it was written
against.

Extracting `createJsonSchema` orphaned nine imports here. They are gone.
